Split part of code from ProductCard into new ProductRating component
To make testing easier - or rather, improving the
design and adhering to the single responsibility
principle a little better.
